+++ main/java/jonatan/stryktipset/Match.java
@@ -3,14 +3,14 @@ package jonatan.stryktipset;
 public class Match
 {
 	private final int id;
-	private final float _odds1;
-	private final float _oddsX;
-	private final float _odds2;
-	private float _probability1;
-	private float _probabilityX;
-	private float _probability2;
+	private final double _odds1;
+	private final double _oddsX;
+	private final double _odds2;
+	private double _probability1;
+	private double _probabilityX;
+	private double _probability2;
 
-	public Match(int match_id, float input1, float inputX, float input2)
+	public Match(int match_id, double input1, double inputX, double input2)
 	{
 		id = match_id;
 		_odds1 = input1;
@@ -25,50 +25,50 @@ public class Match
 		return String.format("Match %3d: %5.2f %5.2f %5.2f", id, _probability1, _probabilityX, _probability2);
 	}
 
-	private void calculateProbabilites(float odds1, float oddsX, float odds2)
+	private void calculateProbabilites(double odds1, double oddsX, double odds2)
 	{
-		float total = 1 / odds1 + 1 / oddsX + 1 / odds2;
-		float factor = 1 / total;
+		double total = 1 / odds1 + 1 / oddsX + 1 / odds2;
+		double factor = 1 / total;
 		_probability1 = factor * 1 / odds1;
 		_probabilityX = factor * 1 / oddsX;
 		_probability2 = factor * 1 / odds2;
-		float totalProbability = _probability1 + _probabilityX + _probability2;
+		double totalProbability = _probability1 + _probabilityX + _probability2;
 		if(totalProbability > 1 + 1e-6 || totalProbability < 1 - 1e-6) {
 			throw new RuntimeException(String.format("Total probability of match %d is not close to 1 ( %f )", id, totalProbability));
 		}
 	}
 
-	public float getProbability1()
+	public double getProbability1()
 	{
 		return _probability1;
 	}
 
-	public float getProbabilityX()
+	public double getProbabilityX()
 	{
 		return _probabilityX;
 	}
 
-	public float getProbability2()
+	public double getProbability2()
 	{
 		return _probability2;
 	}
 
-	public float getOdds1()
+	public double getOdds1()
 	{
 		return _odds1;
 	}
 
-	public float getOddsX()
+	public double getOddsX()
 	{
 		return _oddsX;
 	}
 
-	public float getOdds2()
+	public double getOdds2()
 	{
 		return _odds2;
 	}
 
-	public float getProbability(Result result)
+	public double getProbability(Result result)
 	{
 		switch(result) {
 		case _1:
